Fixed clear cart defect and moved cart routes out of main to the cart module
